[FIX][유종문] preprocess branch 리팩토링
