--- crawler/README.md
@@ -1,16 +1,0 @@
-## ChromeDriver 설치
-
-https://chromedriver.chromium.org/downloads
-
-- 위 링크에서 사용하는 chrome 버전에 맞춰 설치 진행
-
-- `chromedriver.exe`는 `/crawler` 폴더 내부에 위치해야함
-
-- chrome 버전 확인  
-    `chrome://settings/help`
-    에서 확인 가능
-
-## selenium 설치
-pip install selenium
-
-apt install chromium-chromedriver
